+++ CHANGELOG.md
@@ -5,6 +5,19 @@
 *  🤖 Add husky (#3) (407a637b)
 *  🤖 add git-cz for commit (#1) (ee4a3d88)
 
+##### New Features
+
+*  🎸 Add some entities (2ecd8af9)
+*  🎸 change npm run com to npm run commit (4335601e)
+*  🎸 add entities and some relations (9fc1e7e7)
+*  🎸 Add some entities (52b08f1c)
+*  🎸 change npm run com to npm run commit (84072528)
+*  🎸 add entities and some relations (197eb0c0)
+
+##### Bug Fixes
+
+*  🐛 add module attraction (41a3915e)
+
 ##### Documentation Changes
 
 *  ✏️ improve readme (a7ecdf85)
@@ -20,9 +33,14 @@
 
 ##### Other Changes
 
+
+* TesisUNL/PTT_backend into feature/entities (61548191)
+* TesisUNL/PTT_backend (3044ffba)
+
 *  add mvc structure and hbs template engine (8eaebad8)
 
 ##### Refactors
 
 *  💡 Improve project structure and fix packages (#10) (899ff1ae)
+